Make WitnessMap its own module
