+++ acir/src/native_types/mod.rs
@@ -1,6 +1,7 @@
 mod expression;
 mod witness;
+mod witness_map;
 
 pub use expression::Expression;
 pub use witness::Witness;
-pub use witness::WitnessMap;
+pub use witness_map::WitnessMap;
